Revert "[app] 페이지 내비게이터 추가 (#2)"
This reverts commit bc9873b8b0f4ac5901ec5e84e2fbdbfadfc5c9ec.
