+++ .gitignore
@@ -58,6 +58,3 @@ buck-out/
 
 # CocoaPods
 /ios/Pods/
-
-# Jest
-.jest/
