+++ ios/Podfile
@@ -12,7 +12,6 @@ target 'banbanTaxi' do
     :hermes_enabled => false
   )
 
-
   target 'banbanTaxiTests' do
     inherit! :complete
     # Pods for testing
