+++ package.json
@@ -4,57 +4,30 @@
   "private": true,
   "scripts": {
     "android": "react-native run-android",
-    "ios": "npm run pod:install && react-native run-ios",
-    "ios:clean": "rm -rf ~/Library/Developer/Xcode/DerivedData/*",
-    "pod:install": "npm run pod:clean && cd ios && pod install && cd ../",
-    "pod:clean": "cd ios && pod cache clean --all",
+    "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest",
     "lint": "eslint ."
   },
   "dependencies": {
-    "@react-navigation/native": "^6.0.6",
-    "@react-navigation/native-stack": "^6.2.5",
     "react": "17.0.2",
-    "react-native": "0.66.4",
-    "react-native-safe-area-context": "^3.3.2",
-    "react-native-screens": "^3.10.1",
-    "react-native-vector-icons": "^9.0.0"
+    "react-native": "0.66.4"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
     "@babel/runtime": "^7.12.5",
     "@react-native-community/eslint-config": "^2.0.0",
-    "@types/jest": "^27.4.0",
     "@types/react": "^17.0.38",
     "@types/react-native": "^0.66.11",
-    "@types/react-native-vector-icons": "^6.4.10",
-    "@types/react-test-renderer": "^17.0.1",
     "babel-jest": "^26.6.3",
     "eslint": "7.14.0",
     "jest": "^26.6.3",
     "metro-react-native-babel-preset": "^0.66.2",
     "react-native-typescript-transformer": "^1.2.13",
     "react-test-renderer": "17.0.2",
-    "ts-jest": "^27.1.2",
     "typescript": "^4.5.4"
   },
   "jest": {
-    "preset": "react-native",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js"
-    ],
-    "transform": {
-      "^.+\\.(js)$": "<rootDir>/node_modules/babel-jest",
-      "\\.(ts|tsx)$": "<rootDir>/node_modules/ts-jest/preprocessor.js"
-    },
-    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
-    "testPathIgnorePatterns": [
-      "\\.snap$",
-      "<rootDir>/node_modules/"
-    ],
-    "cacheDirectory": ".jest/cache"
+    "preset": "react-native"
   }
 }
