--- src/screens/menus/MenuScreen.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import {Text} from 'react-native';
-
-function MenuScreen() {
-  return <Text>메뉴 스크린</Text>;
-}
-
-export default MenuScreen;
