--- src/shared/constants/NavigatorTypes.ts
@@ -1,8 +1,0 @@
-import {NativeStackScreenProps} from 'react-native-screens/native-stack';
-
-export type RootStackParamList = {
-  Home: undefined;
-  Menu: undefined;
-};
-
-export type HomeProps = NativeStackScreenProps<RootStackParamList, 'Home'>;
